==> 3. wolf analyses/4.  wolf conflict scripts/1. wolf increase data.R <==

# Prep Wolf Public Opinion Data: ------------------------------------------
  ## Here we bring in the wolf increase public response data and crop it to the BH Subwatershed region


# Load packages: ----------------------------------------------------------
library(terra)



# Bring in Data: ----------------------------------------------------------
